Fix import error and move directory
