+++ oslo/torch/nn/parallel/distributed/data_parallel/_param_bucket.py
