+++ tests/nn/data_parallel/test_fsdp_input.py
@@ -21,7 +21,7 @@ from torch.optim import SGD
 from oslo.torch.distributed import ParallelContext
 from oslo.torch.nn.parallel.distributed import FullyShardedDataParallel as FSDP
 from oslo.torch.nn.parallel.distributed import TrainingState
-from oslo.torch.utils import torch_version
+from oslo.torch.utils.version import torch_version
 from oslo.torch.utils.testing import dist_init, rmf, skip_if_no_cuda, teardown
